added about section to homepage
