+++ client/src/components/home/HeroBanner.js
@@ -2,32 +2,6 @@ import React from 'react'
 import { isMobile } from "react-device-detect";
 import { Container } from 'reactstrap';
 
-const styles = {
-  bannerCoverPc: {
-    backgroundImage: 'url("/images/banner-cover-pc.jpg")',
-    backgroundSize: 'cover'
-  },
-  bannerCoverMobile: {
-    backgroundImage: 'url("/images/banner-cover-mobile.jpg")',
-    backgroundSize: 'cover'
-  },
-  textBanner: {
-    textShadow: "3px 3px 3px grey",
-    textAlign: 'center',
-    color:'white'
-  },
-  centerButtons: {
-    textAlign: 'center',
-    padding: '30px'
-  },
-  titleH1Pc: {
-    fontSize: '80px'
-  },
-  titleH1Mobile: {
-    fontSize: '60px'
-  }
-};
-
 const { bannerCoverPc, bannerCoverMobile, textBanner, titleH1Mobile, titleH1Pc } = styles
 
 const HeroBanner = () => (
